requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-aiplatform
numpy
tiktoken
pdfplumber
openai
python-multipart